requirements.txt
asgiref==3.2.3
astroid==2.3.3
captcha==0.3
certifi==2021.5.30
charset-normalizer==2.0.4
Django==3.0.3
django-bootstrap-form==3.4
django-ckeditor==6.1.0
django-crispy-forms==1.12.0
django-extensions==3.1.3
django-filter==21.1
django-js-asset==1.2.2
django-recaptcha==2.0.6
et-xmlfile==1.1.0
idna==3.2
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
openpyxl==3.0.7
Pillow==8.3.2
pylint==2.4.4
python-decouple==3.5
pytz==2019.3
requests==2.26.0
six==1.14.0
slugify==0.0.1
sqlparse==0.3.0
typed-ast==1.4.1
urllib3==1.26.6
whitenoise==5.0.1
wrapt==1.11.2
